Updating Package, abandoning.
As mentioned in [this issue](https://github.com/apollographql/apollo-server/issues/1308), Apollo Server 2.0 doesn't work with the way this repo had been built. This was just an experiment anyway (though with hope to use in real world applications), and considering my other hesitations about how Apollo works, this is just the deciding factor. I'm going to try a different approach.